Some fixes...

+++ Realm/src/main/java/org/jmangos/realm/model/enums/HighGuid.java
@@ -90,7 +90,7 @@ public enum HighGuid {
      * 
      * @return the value
      */
-    public long getValue() {
+    public int getValue() {
     
         return this.value;
     }
